fix varrock collision, start on interfaces

+++ pkg/models/Chunk.go
@@ -8,7 +8,7 @@ type Chunk struct {
 	CollisionMatrix []*CollisionMatrix // a matrix for each height level (4)
 }
 
-func (c *Chunk) ToTile(x, y uint16) *Tile {
+func (c *Chunk) ToTile(x, y int) *Tile {
 	return &Tile{X: (x + 6) << 3, Y: (y + 6) << 3}
 }
